guido_backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-genai==0.6.0
gTTS==2.5.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg[binary]>=3.3.0
PyJWT==2.8.0
bcrypt==4.1.2
